+++ .travis.yml
@@ -3,7 +3,7 @@ os:
   - linux
   - osx
 node_js:
-  - "6"
+  - "8"
 before_install:
 - npm -g install npm@4
 script: npm run ci
