+++ appveyor.yml
@@ -14,7 +14,7 @@ matrix:
 # what combinations to test
 environment:
   matrix:
-    - nodejs_version: 6
+    - nodejs_version: 8
 
 install:
   - ps: Install-Product node $env:nodejs_version x64
