+++ lib/utils/device/deploy.ts
@@ -165,7 +165,7 @@ export async function deployToDevice(opts: DeviceDeployOptions): Promise<void> {
 		});
 
 		globalLogger.logLivepush('Watching for file changes...');
-		const promises = [livepush.init()];
+		const promises: Array<Bluebird<void> | Promise<void>> = [livepush.init()];
 		// Only show logs if we're not detaching
 		if (!opts.detached) {
 			console.log();
