+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "es6",
+		"target": "es2015",
 		"outDir": "build",
 		"strict": true,
 		"strictPropertyInitialization": false,
@@ -12,19 +12,17 @@
 		"sourceMap": true,
 		"skipLibCheck": true,
 		"lib": [
-			// es5 defaults:
-			"dom",
-			"es5",
-			"scripthost",
-			// some specific es6 bits we're sure are safe:
-			"es2015.collection",
-			"es2015.iterable",
-			"es2016.array.include"
+			// Additional API groups compatible with Node 8+ (https://node.green)
+			"es2016.array.include",
+			"es2017.object"
+		],
+		"typeRoots" : [
+			"./node_modules/@types",
+			"./node_modules/etcher-sdk/typings",
+			"./typings"
 		]
 	},
 	"include": [
-		"./typings/*.d.ts",
-		"./node_modules/etcher-sdk/typings/**/*.d.ts",
 		"./lib/**/*.ts"
 	]
 }
