--- typings/@resin-valid-email.d.ts
@@ -1,1 +1,0 @@
-declare module '@resin.io/valid-email';
