--- typings/balena-sync.d.ts
@@ -1,5 +1,0 @@
-declare module 'balena-sync' {
-	import { CommandDefinition } from 'capitano';
-
-	export function capitano(tool: 'balena-cli'): CommandDefinition;
-}
