+++ typings/color-hash/index.d.ts
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright 2019 Balena Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+declare module 'color-hash' {
+	interface Hasher {
+		hex(text: string): string;
+	}
+
+	class ColorHash {
+		public hex(text: string): string;
+		public rgb(text: string): [number, number, number];
+	}
+
+	export = ColorHash;
+}
