--- typings/ent.d.ts
@@ -1,1 +1,0 @@
-declare module 'ent';
