--- typings/global.d.ts
@@ -1,3 +1,0 @@
-interface Dictionary<T> {
-	[key: string]: T;
-}
