--- typings/package.json.d.ts
@@ -1,4 +1,0 @@
-declare module '*/package.json' {
-	export const name: string;
-	export const version: string;
-}
