--- typings/president.d.ts
@@ -1,6 +1,0 @@
-declare module 'president' {
-	export function execute(
-		command: string[],
-		callback: (err: Error) => void,
-	): void;
-}
