--- typings/resin-cli-form.d.ts
@@ -1,1 +1,0 @@
-declare module 'resin-cli-form';
