--- typings/resin-cli-visuals.d.ts
@@ -1,1 +1,0 @@
-declare module 'resin-cli-visuals';
