--- typings/resin-image-fs.d.ts
@@ -1,5 +1,0 @@
-declare module 'resin-image-fs' {
-	import Promise = require('bluebird');
-
-	export function readFile(options: {}): Promise<string>;
-}
