+++ typings/update-notifier/index.d.ts
